site/backend/drf/requirements.txt
asgiref==3.6.0
attrs==22.2.0
colorama==0.4.6
Django==4.1.7
django-cors-headers==4.1.0
django-filter==22.1
djangorestframework==3.2.3
exceptiongroup==1.1.2
iniconfig==2.0.0
Markdown==3.4.1
packaging==23.0
pluggy==1.0.0
pytest==7.2.1
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7